rename commits to logs to make space for conflicts
