+++ program/logs/log/init.go
