+++ program/logs/log/view.go
