+++ program/logs/view.go
@@ -1,4 +1,4 @@
-package commits
+package logs
 
 func (m Model) View() string {
 	return m.list.View()
